+++ src/actions/disable.js
@@ -4,7 +4,9 @@
  */
 
 var SlotFactory = require('../modules/slot-factory.js'),
-    GroupFactory = require('../modules/group-factory.js');
+    GroupFactory = require('../modules/group-factory.js'),
+    events = require('../modules/master-event-handler.js'),
+    googletag = require('../modules/googletag.js');
 
 /**
  * ## harmony.disable
@@ -28,5 +30,27 @@ module.exports = {
         group.forEach(function (slot) {
             slot.enabled = false;
         });
-    }
+    },
+    /**
+     * ## initialLoad([restoreAfterSRM])
+     * @param {?boolean} restoreAfterSRM True to reenable initial load after the
+     * SRM call returns.
+     * Disable initial slot load, but offer the option to reenable. This is a
+     * **massive** oversight by the GPT library. Hopefully one day I can delete
+     * this code.
+     */
+    initialLoad: function (restore) {
+        googletag.pubads().disableInitialLoad();
+        if (restore) {
+            events.one('slotRenderEnded', function () {
+                module.exports.initialLoadRestored = true;
+            });
+        }
+    },
+    /**
+     * ## initialLoadRestored
+     * @type {boolean} True when expecting a single `show` call to display ads
+     * after the SRM ad call has resolved and pubads is set to `disableInitialLoad`.
+     */
+    initialLoadRestored: false
 };