+++ src/modules/master-event-handler.js
@@ -1,0 +1,7 @@
+/**
+ * # Master Event Handler
+ * Handler for system-wide events.
+ */
+
+var EventHandler = require('../types/event-handler.js');
+module.exports = EventHandler();